% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.livereg.R
\name{filter}
\alias{filter}
\title{Filter CSO Live Register data}
\usage{
filter(object, gender = c("both", "male", "female"), age.group = c("all",
  "under_25", "over_and_25"), start.year = 1967, end.year = 2018)
}
\arguments{
\item{object}{\code{livereg} S3 object}

\item{gender}{a character vector of gender choices: "both", "male", "female"}

\item{age.group}{a character vector of age group choices: "all", "under_25", "over_and_25"}

\item{start.year}{the start year of the data}

\item{end.year}{the end year of the data}
}
\value{
A filtered \code{livereg} S3 object which contain the live register unemployment data based on the input filtering parameters.
}
\description{
\code{filter} filters the CSO Live Register unemployment data based on certain input options. These options are gender, age group, start year, and end year.
}
\examples{
dat = load_livereg(use.offline.data = TRUE)
the_1990s_data = filter(dat, start.year = 1990, end.year = 1999)
female_data = filter(dat, gender = c("female"))
}
