% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.livereg.R
\name{fit}
\alias{fit}
\title{Fits a Spline model to the CSO Live Register data}
\usage{
fit(object, spline.type = c("b", "smoothing"), num.knots = 10,
  y.col = NULL)
}
\arguments{
\item{object}{\code{livereg} S3 object}

\item{spline.type}{a character string of spline type choices: "b", "smoothing"}

\item{num.knots}{the number of Spline knots}

\item{y.col}{the name of the column in the data.frame to use as y}
}
\value{
The \code{livereg_fit} S3 object which contains the original data from the \code{livereg} input object as well as the fitted model and predicted value details.
}
\description{
Fits a Spline model to the CSO Live Register unemployment data. The number of knots can be specified so that different model fit can be assessed.
}
\examples{
dat = load_livereg(use.offline.data = TRUE)
fit = fit(dat, "smoothing", num.knots = 5)
}
