% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_livereg.R
\name{load_livereg}
\alias{load_livereg}
\title{Loads the CSO Live Register data}
\usage{
load_livereg(seasonally.adj = TRUE, use.offline.data = TRUE)
}
\arguments{
\item{seasonally.adj}{Whether or not to load the seasonally adjusted live register data}

\item{use.offline.data}{Whether or not to use an offline data file from June 2018}
}
\value{
A \code{livereg} S3 object which contains the live register unemployment data and the timepoints associated with each data value.
}
\description{
Loads the CSO Live Register unemployment data. The data can either be loaded from a offline JSON file (June 2018 data) included in the package or loaded directly from the CSO StatBank API.
}
\examples{
dat = load_livereg(seasonally.adj = TRUE, use.offline.data = TRUE)
}
