% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.livereg_fit.R
\name{plot}
\alias{plot}
\title{Plot the CSO Live Register data and fitted spline model}
\usage{
plot(object, max.x.labels = 10, ...)
}
\arguments{
\item{object}{\code{livereg_fit} S3 object}

\item{max.x.labels}{the maximum number of x axis labels}

\item{...}{additional args}
}
\value{
ggplot of the \code{livereg_fit} S3 object.
}
\description{
Plots the CSO Live Register \code{livereg_fit} S3 object which contains the data and the Spline model fit.
}
\examples{
dat = load_livereg(use.offline.data = TRUE)
fit = fit(dat, spline.type = "smoothing", num.knots = 5)
plot(fit)
}
