% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_livereg.R
\name{shiny_livereg}
\alias{shiny_livereg}
\title{Launch the CSO Umemployment Live Register Shiny App}
\usage{
shiny_livereg()
}
\value{
This method does not return anything. It just launches the Shiny app.
}
\description{
This function launches a Shiny application which uses the functions exported by the \code{liveregr} package.
}
\examples{
\dontrun{
shiny_livereg()
}
}
