% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.livereg_fit.R
\name{summary}
\alias{summary}
\title{Summary of the CSO Live Register data}
\usage{
summary(object)
}
\arguments{
\item{object}{\code{livereg_fit} S3 object}
}
\value{
Summary of the dataset statistics and Spline model of the \code{livereg_fit} object.
}
\description{
Summary of the dataset statistics and Spline model of the CSO Live Register unemployment data.
}
\examples{
dat = load_livereg(use.offline.data = TRUE)
summary(dat)
}
